requirements.txt
# local packages
-e .

hyperopt==0.2.7
imbalanced_learn==0.12.0
joblib==1.3.2
matplotlib==3.8.3
mlflow==2.11.0
numpy==1.26.4
pandas==2.2.1
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.4.1.post1
setuptools==69.0.3
streamlit==1.31.1
